refactor(contact): split contact components
